+++ pkg/embed/embed.go
@@ -4,5 +4,5 @@ package embed
 // Embedder maps texts to dense vectors. Callers pass a batch; the
 // implementation must not create a new client per item.
 type Embedder interface {
-	Embed(texts []string) ([][]float64, error)
+	Embed(texts []string) ([][]float32, error)
 }
